pyproject.toml
[project]
name = "agentic-template"
version = "0.1.0"
description = "Reusable agentic system starter for multi-domain projects"
readme = "README.md"
authors = [{ name = "Agentic Starter" }]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
agentic = "agentic.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
